pyproject.toml
[tool.poetry]
name = "pytube"
version = "0.1.0"
description = ""
authors = ["Paulo <pm85@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
